app/requirements.txt
django==3.1.2
djangorestframework==3.12.1
gunicorn==20.0.4
python-memcached==1.59
